[MOV] test: move some composer tests in correct files
Some tests in the `composer_component.test.ts` were mounting a
spreadsheet after a ComposerWrapper was already mounted in the
beforeEach.

It caused issue with the following commit that adds a Portal target in
the ComposerWrapper. This led to two Portal targets in the document and
a test failing.

This commit moves these tests to the `composer_integration_component.test.ts`
file where they belong.

Task: 4456224
X-original-commit: b1089e3af430246114a1790404da8f0493899392
